+++ main.py
@@ -18,10 +18,6 @@ kivy.require('1.11.0')
 # TODO: Set up options menu
 
 
-def asset_path(file):
-    return os.path.join('assets', file)
-
-
 class MainMenu(Screen):
     pass
 
@@ -44,22 +40,33 @@ class GameScreen(Screen):
 
 class TicTacToeApp(App):
 
+    __sm = None
+
     def config_setup(self):
+        """
+        Configures the program
+        :return:    None
+        """
         path = os.path.dirname(os.path.realpath(__file__))
         os.chdir(path)
         self.title = 'TicTacToePy'
-        self.icon = asset_path('icon.png')
+        self.icon = os.path.join('assets', 'icon.png')
         Window.fullscreen = 'auto'
 
     @staticmethod
     def get_sm():
-        sm = ScreenManager(transition=SlideTransition())
-        sm.add_widget(MainMenu(name='menu'))
-        sm.add_widget(PlayMenu(name='play'))
-        sm.add_widget(SettingsMenu(name='settings'))
-        sm.add_widget(GameScreen(name='sp', game_mode=GameMode.SINGLE_PLAYER))
-        sm.add_widget(GameScreen(name='mp', game_mode=GameMode.MULTI_PLAYER))
-        return sm
+        """
+        Generates the screen manager if it is None and returns it
+        :return:    The program's ScreenManager
+        """
+        if TicTacToeApp.__sm is None:
+            TicTacToeApp.__sm = ScreenManager(transition=SlideTransition())
+            TicTacToeApp.__sm.add_widget(MainMenu(name='menu'))
+            TicTacToeApp.__sm.add_widget(PlayMenu(name='play'))
+            TicTacToeApp.__sm.add_widget(SettingsMenu(name='settings'))
+            TicTacToeApp.__sm.add_widget(GameScreen(name='sp', game_mode=GameMode.SINGLE_PLAYER))
+            TicTacToeApp.__sm.add_widget(GameScreen(name='mp', game_mode=GameMode.MULTI_PLAYER))
+        return TicTacToeApp.__sm
 
     def build(self):
         self.config_setup()